Introduce AiService boundary interface, NoOpAiService, and AiServiceProvider SPI
